+++ cs/unity/Assets/Editor/EnumerateAdaptersWindow.cs
@@ -4,7 +4,7 @@
  * Created Date: 03/07/2020
  * Author: Shun Suzuki
  * -----
- * Last Modified: 10/10/2022
+ * Last Modified: 26/10/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2021 Shun Suzuki. All rights reserved.
@@ -14,6 +14,10 @@
 using UnityEditor;
 using UnityEngine;
 using System.Linq;
+
+#if UNITY_2020_2_OR_NEWER
+#nullable enable
+#endif
 
 public class EnumerateAdaptersWindow : EditorWindow
 {
@@ -53,3 +57,7 @@ public class EnumerateAdaptersWindow : EditorWindow
         }
     }
 }
+
+#if UNITY_2020_2_OR_NEWER
+#nullable disable
+#endif