+++ firmware/fpga/rtl/sources_1/new/pwm/time_cnt_gen.sv
@@ -4,7 +4,7 @@
  * Created Date: 15/03/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 28/07/2022
+ * Last Modified: 31/10/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -19,6 +19,8 @@ module time_cnt_generator #(
     input var CLK,
     input var [63:0] SYS_TIME,
     input var [WIDTH-1:0] CYCLE[0:DEPTH-1],
+    input var [WIDTH-1:0] CYCLE_M1[0:DEPTH-1],
+    input var [WIDTH-1:0] CYCLE_M2[0:DEPTH-1],
     output var [WIDTH-1:0] TIME_CNT[0:DEPTH-1]
 );
 
@@ -34,9 +36,6 @@ module time_cnt_generator #(
   bit [$clog2(DEPTH)-1:0] sync_cnt = DIV_LATENCY % DEPTH;
   bit [$clog2(DEPTH)-1:0] set_cnt = 0;
 
-  bit [WIDTH-1:0] cycle_m1[0:DEPTH];
-  bit [WIDTH-1:0] cycle_m2[0:DEPTH];
-
   div_64_16 div_64_16 (
       .s_axis_dividend_tdata(divined),
       .s_axis_dividend_tvalid(1'b1),
@@ -50,9 +49,9 @@ module time_cnt_generator #(
   for (genvar i = 0; i < DEPTH; i++) begin
     always_ff @(posedge CLK) begin
       if (i == set_cnt) begin
-        t[i] <= (t[i] == cycle_m2[i]) && (rem[WIDTH-1:0] == 0) ? t[i] + 1 : rem[WIDTH-1:0]; // make sure t be T-1
+        t[i] <= (t[i] == CYCLE_M2[i]) && (rem[WIDTH-1:0] == 0) ? t[i] + 1 : rem[WIDTH-1:0]; // make sure t be T-1
       end else begin
-        t[i] <= (t[i] == cycle_m1[i]) ? 0 : t[i] + 1;
+        t[i] <= (t[i] == CYCLE_M1[i]) ? 0 : t[i] + 1;
       end
     end
     assign TIME_CNT[i] = t[i];
@@ -66,11 +65,5 @@ module time_cnt_generator #(
     set_cnt  <= (set_cnt == DEPTH - 1) ? 0 : set_cnt + 1;
   end
 
-  for (genvar i = 0; i < DEPTH; i++) begin
-    always_ff @(posedge CLK) begin
-      cycle_m1[i] <= CYCLE[i] - 1;
-      cycle_m2[i] <= cycle_m1[i] - 1;
-    end
-  end
 
 endmodule