+++ julia/src/SOEM.jl
@@ -3,7 +3,7 @@
 # Created Date: 13/06/2022
 # Author: Shun Suzuki
 # -----
-# Last Modified: 10/10/2022
+# Last Modified: 04/11/2022
 # Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 # -----
 # Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -11,21 +11,21 @@
 
 struct SOEM
     _link::Link
-    function SOEM(; ifname::String="", send_cycle=1, sync0_cycle=1, freerun::Bool=false, on_lost=Nothing, high_precision::Bool=false)
+    function SOEM(; ifname::String="", send_cycle=2, sync0_cycle=2, freerun::Bool=false, on_lost=Nothing, high_precision::Bool=false, check_interval=500)
         chandle = Ref(Ptr{Cvoid}(0))
         if on_lost == Nothing
             if ifname == ""
-                autd3capi_link_soem.autd_link_soem(chandle, Ptr{Cvoid}(C_NULL), UInt16(sync0_cycle), UInt16(send_cycle), freerun, Nothing, high_precision)
+                autd3capi_link_soem.autd_link_soem(chandle, Ptr{Cvoid}(C_NULL), UInt16(sync0_cycle), UInt16(send_cycle), freerun, Nothing, high_precision, UInt64(check_interval))
             else
-                autd3capi_link_soem.autd_link_soem(chandle, ifname, UInt16(sync0_cycle), UInt16(send_cycle), freerun, Nothing, high_precision)
+                autd3capi_link_soem.autd_link_soem(chandle, ifname, UInt16(sync0_cycle), UInt16(send_cycle), freerun, Nothing, high_precision, UInt64(check_interval))
             end
         else
             f = (x::Cstring) -> on_lost(x)
             p = @cfunction($f, Cvoid, (Cstring,))
             if ifname == ""
-                autd3capi_link_soem.autd_link_soem(chandle, Ptr{Cvoid}(C_NULL), UInt16(sync0_cycle), UInt16(send_cycle), freerun, p, high_precision)
+                autd3capi_link_soem.autd_link_soem(chandle, Ptr{Cvoid}(C_NULL), UInt16(sync0_cycle), UInt16(send_cycle), freerun, p, high_precision, UInt64(check_interval))
             else
-                autd3capi_link_soem.autd_link_soem(chandle, ifname, UInt16(sync0_cycle), UInt16(send_cycle), freerun, p, high_precision)
+                autd3capi_link_soem.autd_link_soem(chandle, ifname, UInt16(sync0_cycle), UInt16(send_cycle), freerun, p, high_precision, UInt64(check_interval))
             end
         end
         new(Link(chandle[]))