+++ nim/src/autd3/native_methods/autd3capi_link_simulator.h
@@ -12,4 +12,4 @@
 #  endif
 #endif
 
-void AUTDLinkSimulator(void** out, uint16 port, char* ip_addr);
+void AUTDLinkSimulator(void** out);