+++ nim/src/autd3/native_methods/autd3capi_link_soem.nim
@@ -18,4 +18,9 @@ proc AUTDFreeAdapterPointer*(p_adapter: pointer) {.cdecl,
     importc: "AUTDFreeAdapterPointer", dynlib: dll.}
 proc AUTDLinkSOEM*(`out`: ptr pointer; ifname: cstring; sync0_cycle: uint16;
                   send_cycle: uint16; freerun: bool; on_lost: pointer;
-                  high_precision: bool) {.cdecl, importc: "AUTDLinkSOEM", dynlib: dll.}
+                  high_precision: bool; state_check_interval: uint64) {.cdecl,
+    importc: "AUTDLinkSOEM", dynlib: dll.}
+proc AUTDLinkSOEMSetLogLevel*(level: int32) {.cdecl,
+    importc: "AUTDLinkSOEMSetLogLevel", dynlib: dll.}
+proc AUTDLinkSOEMSetDefaultLogger*(`out`: pointer; flush: pointer) {.cdecl,
+    importc: "AUTDLinkSOEMSetDefaultLogger", dynlib: dll.}